# docs/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: NENE2 API
  version: 0.1.0
  description: >
    NENE2 API contract. This file is the source of truth for public JSON APIs.
servers:
  - url: http://localhost:8080
    description: Local Docker development server
tags:
  - name: System
    description: Framework and runtime smoke endpoints.
  - name: Exhibitions
    description: Public SAKURA Exhibition-style field-trial endpoints.
paths:
  /:
    get:
      tags:
        - System
      summary: Framework smoke endpoint
      description: Returns a minimal JSON response that confirms the NENE2 public entry point is working.
      operationId: getFrameworkSmoke
      responses:
        '200':
          description: Framework smoke response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FrameworkSmokeResponse'
              examples:
                ok:
                  summary: Successful smoke response
                  value:
                    name: NENE2
                    description: JSON APIs first, minimal server HTML, frontend ready, AI-readable.
                    status: ok
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /health:
    get:
      tags:
        - System
      summary: Health endpoint
      description: Returns a minimal JSON response for local operational health checks.
      operationId: getHealth
      responses:
        '200':
          description: Health response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'
              examples:
                ok:
                  summary: Healthy response
                  value:
                    status: ok
                    service: NENE2
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /machine/health:
    get:
      tags:
        - System
      summary: Protected machine health endpoint
      description: Returns health information for machine clients that provide a configured API key.
      operationId: getMachineHealth
      security:
        - ApiKeyAuth:
            - read:health
      responses:
        '200':
          description: Machine health response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MachineHealthResponse'
              examples:
                ok:
                  summary: Authenticated machine health response
                  value:
                    status: ok
                    service: NENE2
                    credential_type: api_key
        '401':
          $ref: '#/components/responses/Unauthorized'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /examples/ping:
    get:
      tags:
        - System
      summary: Example ping endpoint
      description: Returns a minimal response used to exercise the endpoint scaffold workflow.
      operationId: getExamplePing
      responses:
        '200':
          description: Example ping response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExamplePingResponse'
              examples:
                ok:
                  summary: Successful example ping response
                  value:
                    message: pong
                    status: ok
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /exhibitions/{year}/artists:
    get:
      tags:
        - Exhibitions
      summary: Exhibition artist list by year
      description: Returns the fixed sandbox artist list for a SAKURA Exhibition-style field-trial year.
      operationId: getExhibitionArtistsByYear
      parameters:
        - name: year
          in: path
          required: true
          description: Exhibition year to read from the sandbox fixture.
          schema:
            type: integer
            format: int32
            minimum: 2009
            maximum: 2026
          example: 2025
      responses:
        '200':
          description: Exhibition artist list response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExhibitionArtistsResponse'
              examples:
                ok:
                  summary: 2025 artist list
                  value:
                    exhibitionYear: 2025
                    artists:
                      - artistId: 1
                        displayName:
                          en: Yoshimi Ohtani
                          jp: オオタニヨシミ
                        countryOrRegion: Japan
                        workCount: 2
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /exhibitions/{year}/works:
    get:
      tags:
        - Exhibitions
      summary: Exhibition work list by year
      description: Returns the fixed sandbox work list for a SAKURA Exhibition-style field-trial year.
      operationId: getExhibitionWorksByYear
      parameters:
        - name: year
          in: path
          required: true
          description: Exhibition year to read from the sandbox fixture.
          schema:
            type: integer
            format: int32
            minimum: 2009
            maximum: 2026
          example: 2026
      responses:
        '200':
          description: Exhibition work list response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExhibitionWorksResponse'
              examples:
                ok:
                  summary: 2026 work list
                  value:
                    exhibitionYear: 2026
                    works:
                      - workId: 20260101
                        artistId: 1
                        artistDisplayName:
                          en: Yoshimi Ohtani
                          jp: オオタニヨシミ
                        title:
                          en: Spring Light
                          jp: 春の光
                        workNumber: 1
                      - workId: 20260201
                        artistId: 2
                        artistDisplayName:
                          en: Hideyuki Mori
                          jp: 彩
                        title:
                          en: Color Field
                          jp: 彩の場
                        workNumber: 1
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /exhibitions/{year}/works/{workId}:
    get:
      tags:
        - Exhibitions
      summary: Exhibition work detail by year and id
      description: >-
        Returns a single sandbox work for the exhibition year when it exists.
        Returns 404 when the work id is not registered for that year.
      operationId: getExhibitionWorkByYearAndId
      parameters:
        - name: year
          in: path
          required: true
          description: Exhibition year to read from the sandbox fixture.
          schema:
            type: integer
            format: int32
            minimum: 2009
            maximum: 2026
          example: 2026
        - name: workId
          in: path
          required: true
          description: Sandbox work identifier.
          schema:
            type: integer
            format: int32
          example: 20260101
      responses:
        '200':
          description: Exhibition work detail response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExhibitionWorkDetailResponse'
              examples:
                ok:
                  summary: 2026 work detail
                  value:
                    exhibitionYear: 2026
                    work:
                      workId: 20260101
                      artistId: 1
                      artistDisplayName:
                        en: Yoshimi Ohtani
                        jp: オオタニヨシミ
                      title:
                        en: Spring Light
                        jp: 春の光
                      workNumber: 1
        '404':
          $ref: '#/components/responses/NotFound'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /exhibitions/2026/artists:
    get:
      tags:
        - Exhibitions
      summary: 2026 exhibition artist list
      description: Returns the fixed sandbox artist list for the 2026 SAKURA Exhibition-style field trial.
      operationId: getExhibition2026Artists
      responses:
        '200':
          description: Exhibition artist list response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExhibitionArtistsResponse'
              examples:
                ok:
                  summary: 2026 artist list
                  value:
                    exhibitionYear: 2026
                    artists:
                      - artistId: 1
                        displayName:
                          en: Yoshimi Ohtani
                          jp: オオタニヨシミ
                        countryOrRegion: Japan
                        workCount: 1
                      - artistId: 2
                        displayName:
                          en: Hideyuki Mori
                          jp: 彩
                        countryOrRegion: Japan
                        workCount: 1
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /exhibitions/2026/works:
    get:
      tags:
        - Exhibitions
      summary: 2026 exhibition work list
      description: Returns the fixed sandbox work list for the 2026 SAKURA Exhibition-style field trial.
      operationId: getExhibition2026Works
      responses:
        '200':
          description: Exhibition work list response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExhibitionWorksResponse'
              examples:
                ok:
                  summary: 2026 work list
                  value:
                    exhibitionYear: 2026
                    works:
                      - workId: 20260101
                        artistId: 1
                        artistDisplayName:
                          en: Yoshimi Ohtani
                          jp: オオタニヨシミ
                        title:
                          en: Spring Light
                          jp: 春の光
                        workNumber: 1
                      - workId: 20260201
                        artistId: 2
                        artistDisplayName:
                          en: Hideyuki Mori
                          jp: 彩
                        title:
                          en: Color Field
                          jp: 彩の場
                        workNumber: 1
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '500':
          $ref: '#/components/responses/InternalServerError'
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-NENE2-API-Key
      description: API key for machine clients. Do not commit real key values.
  responses:
    Unauthorized:
      description: A valid API key is required for this endpoint.
      headers:
        WWW-Authenticate:
          description: API-key challenge for machine clients.
          schema:
            type: string
            example: ApiKey realm="NENE2", header="X-NENE2-API-Key"
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
          examples:
            unauthorized:
              summary: Missing or invalid API key
              value:
                type: https://nene2.dev/problems/unauthorized
                title: Unauthorized
                status: 401
                detail: A valid API key is required for this endpoint.
                instance: /machine/health
    NotFound:
      description: The requested resource was not found.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
          examples:
            notFound:
              summary: Resource not found
              value:
                type: https://nene2.dev/problems/not-found
                title: Not Found
                status: 404
                detail: The requested resource was not found.
                instance: /missing
    MethodNotAllowed:
      description: The requested resource does not support this HTTP method.
      headers:
        Allow:
          description: Comma-separated list of allowed HTTP methods.
          schema:
            type: string
            example: GET
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
          examples:
            methodNotAllowed:
              summary: Method not allowed
              value:
                type: https://nene2.dev/problems/method-not-allowed
                title: Method Not Allowed
                status: 405
                detail: The requested resource does not support this HTTP method.
                instance: /
    ValidationFailed:
      description: The request body, query, or path parameters contain invalid values.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ValidationProblemDetails'
          examples:
            validationFailed:
              summary: Validation failed
              value:
                type: https://nene2.dev/problems/validation-failed
                title: Validation Failed
                status: 422
                detail: The request body contains invalid values.
                errors:
                  - field: email
                    message: Email must be a valid email address.
                    code: invalid_email
    PayloadTooLarge:
      description: The request body exceeds the configured size limit.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
          examples:
            payloadTooLarge:
              summary: Payload too large
              value:
                type: https://nene2.dev/problems/payload-too-large
                title: Payload Too Large
                status: 413
                detail: The request body exceeds the configured size limit.
                instance: /upload
                max_body_bytes: 1048576
    InternalServerError:
      description: The server encountered an unexpected condition.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
          examples:
            internalServerError:
              summary: Internal server error
              value:
                type: https://nene2.dev/problems/internal-server-error
                title: Internal Server Error
                status: 500
                detail: The server encountered an unexpected condition.
  schemas:
    ProblemDetails:
      type: object
      description: RFC 9457 Problem Details response for NENE2 JSON APIs.
      required:
        - type
        - title
        - status
      properties:
        type:
          type: string
          format: uri
          description: Stable problem type URI.
          example: https://nene2.dev/problems/not-found
        title:
          type: string
          description: Short, human-readable summary of the problem type.
          example: Not Found
        status:
          type: integer
          format: int32
          minimum: 400
          maximum: 599
          description: HTTP status code for this occurrence.
          example: 404
        detail:
          type: string
          description: Safe, human-readable detail for this occurrence.
          example: The requested resource was not found.
        instance:
          type: string
          description: Request-specific URI or identifier for this occurrence.
          example: /missing
      additionalProperties: true
    ValidationProblemDetails:
      allOf:
        - $ref: '#/components/schemas/ProblemDetails'
        - type: object
          description: Problem Details response for request validation failures.
          required:
            - errors
          properties:
            errors:
              type: array
              minItems: 1
              items:
                $ref: '#/components/schemas/ValidationError'
    ValidationError:
      type: object
      description: A single request validation error.
      required:
        - field
        - message
        - code
      properties:
        field:
          type: string
          description: Field name, parameter name, or JSON pointer.
          example: email
        message:
          type: string
          description: Safe, human-readable validation message.
          example: Email must be a valid email address.
        code:
          type: string
          description: Stable machine-readable validation code.
          example: invalid_email
      additionalProperties: false
    FrameworkSmokeResponse:
      type: object
      required:
        - name
        - description
        - status
      properties:
        name:
          type: string
          example: NENE2
        description:
          type: string
          example: JSON APIs first, minimal server HTML, frontend ready, AI-readable.
        status:
          type: string
          enum:
            - ok
          example: ok
    HealthResponse:
      type: object
      required:
        - status
        - service
      properties:
        status:
          type: string
          enum:
            - ok
          example: ok
        service:
          type: string
          example: NENE2
    MachineHealthResponse:
      type: object
      required:
        - status
        - service
        - credential_type
      properties:
        status:
          type: string
          enum:
            - ok
          example: ok
        service:
          type: string
          example: NENE2
        credential_type:
          type: string
          enum:
            - api_key
          example: api_key
    ExamplePingResponse:
      type: object
      required:
        - message
        - status
      properties:
        message:
          type: string
          enum:
            - pong
          example: pong
        status:
          type: string
          enum:
            - ok
          example: ok
    ExhibitionArtistsResponse:
      type: object
      required:
        - exhibitionYear
        - artists
      properties:
        exhibitionYear:
          type: integer
          format: int32
          example: 2026
        artists:
          type: array
          items:
            $ref: '#/components/schemas/ExhibitionArtist'
    ExhibitionArtist:
      type: object
      required:
        - artistId
        - displayName
        - countryOrRegion
        - workCount
      properties:
        artistId:
          type: integer
          format: int32
          example: 1
        displayName:
          $ref: '#/components/schemas/ArtistDisplayName'
        countryOrRegion:
          type: string
          example: Japan
        workCount:
          type: integer
          format: int32
          minimum: 0
          example: 1
      additionalProperties: false
    ArtistDisplayName:
      type: object
      required:
        - en
        - jp
      properties:
        en:
          type: string
          example: Yoshimi Ohtani
        jp:
          type: string
          example: オオタニヨシミ
      additionalProperties: false
    ExhibitionWorksResponse:
      type: object
      required:
        - exhibitionYear
        - works
      properties:
        exhibitionYear:
          type: integer
          format: int32
          example: 2026
        works:
          type: array
          items:
            $ref: '#/components/schemas/ExhibitionWork'
    ExhibitionWorkDetailResponse:
      type: object
      required:
        - exhibitionYear
        - work
      properties:
        exhibitionYear:
          type: integer
          format: int32
          example: 2026
        work:
          $ref: '#/components/schemas/ExhibitionWork'
      additionalProperties: false
    ExhibitionWork:
      type: object
      required:
        - workId
        - artistId
        - artistDisplayName
        - title
        - workNumber
      properties:
        workId:
          type: integer
          format: int32
          example: 20260101
        artistId:
          type: integer
          format: int32
          example: 1
        artistDisplayName:
          $ref: '#/components/schemas/ArtistDisplayName'
        title:
          $ref: '#/components/schemas/WorkTitle'
        workNumber:
          type: integer
          format: int32
          minimum: 1
          example: 1
      additionalProperties: false
    WorkTitle:
      type: object
      required:
        - en
        - jp
      properties:
        en:
          type: string
          example: Spring Light
        jp:
          type: string
          example: 春の光
      additionalProperties: false